+++ temba/campaigns/views.py
@@ -330,10 +330,9 @@ class CampaignEventForm(forms.ModelForm):
 
         # if its a message flow, set that accordingly
         if self.cleaned_data["event_type"] == CampaignEvent.TYPE_MESSAGE:
-            if self.instance.id:
-                base_language = self.instance.flow.base_language
-            else:
-                base_language = org.flow_languages[0]
+            base_language = org.flow_languages[0]
+            if self.instance.id and self.instance.base_language:
+                base_language = self.instance.base_language
 
             translations = {}
             for language in self.languages:
@@ -341,13 +340,6 @@ class CampaignEventForm(forms.ModelForm):
                 if iso_code in self.cleaned_data and self.cleaned_data.get(iso_code, "").strip():
                     translations[iso_code] = self.cleaned_data.get(iso_code, "").strip()
 
-            if not obj.flow_id or not obj.flow.is_active or not obj.flow.is_system:
-                obj.flow = Flow.create_single_message(org, request.user, translations, base_language=base_language)
-            else:
-                # set our single message on our flow
-                obj.flow.update_single_message_flow(request.user, translations, base_language)
-
-            obj.message = translations  # deprecated
             obj.translations = {lang: {"text": text} for lang, text in translations.items()}
             obj.base_language = base_language
             obj.full_clean()
@@ -380,13 +372,12 @@ class CampaignEventForm(forms.ModelForm):
 
         if (
             self.instance.id
-            and self.instance.flow
+            and self.instance.event_type == CampaignEvent.TYPE_FLOW
             and self.instance.flow.flow_type == Flow.TYPE_BACKGROUND
-            and not self.instance.message
         ):
             flow.widget.attrs["info_text"] = CampaignEventCRUDL.BACKGROUND_WARNING
 
-        message = self.instance.message or {}
+        message = {lang: t["text"] for lang, t in (self.instance.translations or {}).items()}
         self.languages = []
 
         # add in all of our languages for message forms
@@ -434,9 +425,9 @@ class CampaignEventForm(forms.ModelForm):
         # determine our base language if necessary
         base_language = org.flow_languages[0]
 
-        # if we are editing, always include the flow base language
+        # if we are editing, always include the base language
         if self.instance.id:
-            base_language = self.instance.flow.base_language
+            base_language = self.instance.base_language
 
         # add our default language, we'll insert it at the front of the list
         if base_language and base_language not in self.fields:
@@ -460,7 +451,17 @@ class CampaignEventForm(forms.ModelForm):
 
     class Meta:
         model = CampaignEvent
-        fields = "__all__"
+        fields = (
+            "event_type",
+            "relative_to",
+            "offset",
+            "unit",
+            "delivery_hour",
+            "direction",
+            "flow_to_start",
+            "flow_start_mode",
+            "message_start_mode",
+        )
         widgets = {"offset": InputWidget(attrs={"widget_only": True})}
 
 
@@ -581,10 +582,8 @@ class CampaignEventCRUDL(SmartCRUDL):
             org = self.request.org
             fields += org.flow_languages
 
-            flow_language = self.object.flow.base_language
-
-            if flow_language not in fields:
-                fields.append(flow_language)
+            if self.object.base_language not in fields:
+                fields.append(self.object.base_language)
 
             return fields
 
@@ -605,11 +604,6 @@ class CampaignEventCRUDL(SmartCRUDL):
 
             return initial
 
-        def post_save(self, obj):
-            obj = super().post_save(obj)
-            obj.update_flow_name()
-            return obj
-
         def pre_save(self, obj):
             obj = super().pre_save(obj)
             self.form.pre_save(self.request, obj)
@@ -623,7 +617,6 @@ class CampaignEventCRUDL(SmartCRUDL):
                 flow.save(update_fields=("is_active",))
                 obj.translations = None
                 obj.base_language = None
-                obj.message = None  # deprecated
 
             # if we changed scheduling, update the fire version to invalidate existing fires
             if (
@@ -705,7 +698,6 @@ class CampaignEventCRUDL(SmartCRUDL):
 
         def post_save(self, obj):
             obj = super().post_save(obj)
-            obj.update_flow_name()
             obj.schedule_async()
             return obj