+++ temba/orgs/views/tests/test_orgcrudl.py
@@ -60,7 +60,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
                 "Export",
                 "Import",
                 ("Channels", ["New Channel", "Test Channel"]),
-                ("Classifiers", ["New Classifier"]),
                 ("Archives", ["Messages", "Flow Runs"]),
             ],
             choose_org=self.org,
@@ -135,7 +134,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
                 "Export",
                 "Import",
                 ("Channels", ["New Channel", "Test Channel"]),
-                ("Classifiers", ["New Classifier"]),
                 ("Archives", ["Messages", "Flow Runs"]),
             ],
         )
@@ -176,7 +174,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
                 "Export",
                 "Import",
                 ("Channels", ["New Channel", "Test Channel"]),
-                ("Classifiers", ["New Classifier"]),
                 ("Archives", ["Messages", "Flow Runs"]),
             ],
         )